Corrigido o tipo de retorno em Evento.getChaveImagem() para long

+++ app/src/main/java/com/infact/nightour/model/Evento.java
@@ -124,7 +124,7 @@ public class Evento {
         this.imagem = imagem;
     }
 
-    public int getChaveImagem() {
+    public long getChaveImagem() {
         return getImagem().getId();
     }
